requirements.txt
Django~=3.2.10
earthengine-api
django-crispy-forms==1.14.0
django_debug_toolbar==4.1.0
django-extensions==3.2.3
django-location-field==2.7.2
django-storages==1.13.1
dj-database-url>=0.5,<0.6
crispy-bootstrap5==0.7
folium==0.14.0
gunicorn==20.1.0
Pillow==9.3.0
whitenoise>=6.0,<7.0
psycopg2>=2.0,<3.0